# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-SRO",
    "summary": "Workflow Builder: demonstrations in, reviewable skills out.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/recordings": {
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Start Recording",
        "operationId": "start_recording_v1_recordings_post",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartRecordingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartRecordingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "List Recordings",
        "operationId": "list_recordings_v1_recordings_get",
        "parameters": [
          {
            "name": "objective_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Objective Type"
            }
          },
          {
            "name": "target_system",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target System"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "facility",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Direction"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Direction"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecordingSummary"
                  },
                  "title": "Response List Recordings V1 Recordings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recordings/{recording_id}": {
      "get": {
        "tags": [
          "recordings"
        ],
        "summary": "Get Recording",
        "operationId": "get_recording_v1_recordings__recording_id__get",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recordings/{recording_id}/artifacts": {
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Attach Artifact",
        "operationId": "attach_artifact_v1_recordings__recording_id__artifacts_post",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_attach_artifact_v1_recordings__recording_id__artifacts_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/recordings/{recording_id}/finish": {
      "post": {
        "tags": [
          "recordings"
        ],
        "summary": "Finish Recording",
        "operationId": "finish_recording_v1_recordings__recording_id__finish_post",
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recording Id"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev",
              "title": "X-Tenant-Id"
            }
          },
          {
            "name": "X-Principal-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "dev-operator",
              "title": "X-Principal-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishRecordingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordingSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ArtifactKind": {
        "type": "string",
        "enum": [
          "raw_events",
          "video",
          "screenshot",
          "audio",
          "transcript",
          "payload"
        ],
        "title": "ArtifactKind"
      },
      "ArtifactModel": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "uri": {
            "type": "string",
            "title": "Uri"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "frame_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Index"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "uri",
          "content_type",
          "size_bytes",
          "frame_index",
          "label"
        ],
        "title": "ArtifactModel"
      },
...
      "Body_attach_artifact_v1_recordings__recording_id__artifacts_post": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ArtifactKind"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "file"
        ],
        "title": "Body_attach_artifact_v1_recordings__recording_id__artifacts_post"
      },
      "Direction": {
        "type": "string",
        "enum": [
          "inbound",
          "outbound",
          "internal"
        ],
        "title": "Direction"
      },
      "FinishRecordingRequest": {
        "properties": {
          "abandon_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abandon Reason",
            "description": "Present means abandon; absent means seal."
          }
        },
        "type": "object",
        "title": "FinishRecordingRequest"
      },
      "FrameSummary": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "action_kind": {
            "type": "string",
            "title": "Action Kind"
          },
          "target": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "request_count": {
            "type": "integer",
            "title": "Request Count"
          },
          "primary_request": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Request"
          },
          "error_count": {
            "type": "integer",
            "title": "Error Count"
          }
        },
        "type": "object",
        "required": [
          "index",
          "occurred_at",
          "action_kind",
          "target",
          "request_count",
          "primary_request",
          "error_count"
        ],
        "title": "FrameSummary",
        "description": "Enough to render the timeline. The full frame is fetched per frame."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ObjectiveKeyModel": {
        "properties": {
          "objective_type": {
            "type": "string",
            "title": "Objective Type"
          },
          "target_system": {
            "type": "string",
            "title": "Target System"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "facility": {
            "type": "string",
            "title": "Facility"
          },
          "direction": {
            "$ref": "#/components/schemas/Direction"
          }
        },
        "type": "object",
        "required": [
          "objective_type",
          "target_system",
          "entity_type",
          "facility",
          "direction"
        ],
        "title": "ObjectiveKeyModel"
      },
...
      "RecordingDetail": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "objective_key": {
            "$ref": "#/components/schemas/ObjectiveKeyModel"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "demonstrator": {
            "type": "string",
            "title": "Demonstrator"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "frame_count": {
            "type": "integer",
            "title": "Frame Count"
          },
          "has_narration": {
            "type": "boolean",
            "title": "Has Narration"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/FrameSummary"
            },
            "type": "array",
            "title": "Frames"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactModel"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "id",
          "objective_key",
          "label",
          "status",
          "demonstrator",
          "started_at",
          "ended_at",
          "frame_count",
          "has_narration",
          "frames",
          "artifacts"
        ],
        "title": "RecordingDetail"
      },
      "RecordingSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "objective_key": {
            "$ref": "#/components/schemas/ObjectiveKeyModel"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "demonstrator": {
            "type": "string",
            "title": "Demonstrator"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "frame_count": {
            "type": "integer",
            "title": "Frame Count"
          },
          "has_narration": {
            "type": "boolean",
            "title": "Has Narration"
          }
        },
        "type": "object",
        "required": [
          "id",
          "objective_key",
          "label",
          "status",
          "demonstrator",
          "started_at",
          "ended_at",
          "frame_count",
          "has_narration"
        ],
        "title": "RecordingSummary"
      },
...
      "StartRecordingRequest": {
        "properties": {
          "objective_key": {
            "$ref": "#/components/schemas/ObjectiveKeyModel"
          },
          "start_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Url"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "objective_key"
        ],
        "title": "StartRecordingRequest"
      },
      "StartRecordingResponse": {
        "properties": {
          "recording_id": {
            "type": "string",
            "title": "Recording Id"
          },
          "live_view_url": {
            "type": "string",
            "title": "Live View Url"
          }
        },
        "type": "object",
        "required": [
          "recording_id",
          "live_view_url"
        ],
        "title": "StartRecordingResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}